backend/requirements.txt
Django
djangorestframework
django-cors-headers
django-rest-swagger
PyYAML